[FIX] account: translate names of extra accounts when installing CoA
When we install a chart of accounts (CoA) some extra accounts are added.
These extra accounts do not appear in the chart template CSV file.
Currently no translations are installed for the names of these account.

This commit installs translations for the names of these accounts for all
languages that are installed at the time of the CoA installation.

The problem currently concerns the following accounts
    * (Some) Bank utility accounts
      * Bank Suspense Account
      * Outstanding Receipts
      * Outstanding Payments
      * Cash Difference Loss
      * Cash Difference Gain
      * Liquidity Transfer
    * Bank / Cash journal default accounts
      * Bank
      * Cash
    * Unaffected earnings account
      * Undistributed Profits/Losses

Reproduce:
  1. Make sure the French language is installed but that English is
     still the active language.
     (Settings App > General Settings > Languages)
  2. Install a localisation (e.g. "France - Accounting" / l10n_fr).
  3. Go to the chart of accounts
     (Accounting > Configuration (Menu) > Chart of Accounts)
  4. Switch to the other language on the current user
     (User (on top right) > My profile > Language)
  5. E.g. "Bank Suspense Account" still exists (untranslated)

task-3414329