[IMP] account: make currency_rate of account.move.line stored
This i.e. enables the use of this field in SQL queries.
Previously computations of this rate in SQL queries were either only
approximations or inefficient.

Thus it can i.e. be used to efficiently compute the total of the line
including taxes in company currency (from the total in move currency).
This is needed to display that value in the account invoice report
(see other commits of this PR / task)

task-3613358