[FIX] account: "Total" in company currency on invoice analysis
Currently the column "Total in Currency" displays the value in the currency of the move.
This has issues in multi-currency environments (see below).
This commit renames the column to "Total" and displays the value in
the currency of the active company (after doing the proper
conversions; see below).
A previous try to do this was reverted due to performance issues.

Current multi-currency issue:
Amounts of different currencies may be added together. This yields nonsensical values.
As a workaround it is currently possible to group the values by "Currency" to avoid summing the values (except for the totals).

Changes in this commit:
This commit renames the column to "Total" and displays the amount in the currency of the active company.
The conversion works as follows:
  1. Convert the amount from move currency to move company currency.
     We use the rate on the day of the move for this.
  2. Convert the amount from point 1 to active company currency.
     We use the rate of today of the move for this.

In the previous commit the currency rate of the individual account move lines was made into a stored field.
It is now used here to efficienttly compute step 1 (see above).

task-3613358