[IMP] account: store and display currency rate on moves
Consider an invoice (incl. credit notes, bills etc) in foreign currency.
Before this commit the situation is the following:
  1. Different lines of an invoice may use a different exchange rate.
     This should not be possible.
     But it can e.g. happen when an automatic exchange rate update happens between the creation of the invoice lines.
  2. The exchange rate of an invoice needs to be displayed / exported / reported in some cases (for legal requirements).
     For this the rate needs to be recomputed. This is not really possible due to the following reasons:
       a. There is not really "the" exchange rate of an invoice due to (1.).
       b. The recomputation of the rate may be inexact due to rounding of the stored values.
       c. Different lines may show different exchange rates although the same rate has been used due to (1.) and/or (2.b.).
  3. Some modules already add a similar stored field for the exchange rate to satisfy (2.) and avoid the problems mentioned there.

Thus we add a new stored field 'currency_rate' on model 'account.move' that stores the exchange rate for invoices.
Multiplying a value in company currency with this value yields the value in document currency.
It is displayed on the account.move form view to make it transparent for the user.
This solves problems (2.) and makes the fields from (3.) obsolete.
(See other commits of this PR and the related enterprise PR).
The invoice lines of an invoice will all use the exchange rate of the invoice to solve problem (1.)

Point (2.a.) is also a problem for upgrade:
For pre-existing invoices we use the exchange rate at the date of the receivable / payable line.
(See the related upgrade PR for more details.)

task-3634224