[IMP] account,*: store and display currency rate on moves
*: l10n_ar,l10n_dk_bookkeeping
[REM] l10n_dk_bookkeeping: obsolete since rate is always stored and displayed now

Consider an invoice (incl. credit notes, bills etc) in foreign currency.
Before this commit the situation is the following:
  1. Different lines of an invoice may use a different exchange rate.
     This should not be possible.
     But it can e.g. happen when an automatic exchange rate update happens between the creation of the invoice lines.
  2. The exchange rate of an invoice needs to be displayed in some cases (for legal requirements).
     This is i.e. a problem due to (1.).
  3. Some modules already add a similar stored field for the exchange rate to satisfy (2.) and avoid the problems mentioned there.

Thus we add a new stored field 'currency_rate' on model 'account.move'
that stores the exchange rate (on the invoice date) for invoices.
Multiplying a value in company currency with this value yields the value in document currency.
It is displayed on the account.move form view to make it transparent for the user.
This solves problems (2.) and makes the fields from (3.) obsolete.
The invoice lines of an invoice will all use the exchange rate of the invoice to solve problem (1.)

Point (1.) is also a problem for upgrade:
For pre-existing invoices we use the exchange rate at the date of the receivable / payable line.
(See the related upgrade PR for more details.)

task-3634224