[IMP] l10n_it: add support for Declaration of Intent
In Italy, an operator who exports more than 10% of all his revenues is considered a regular exporter.
Regular exporters can purchase without paying VAT within certain limits.

For this they have to do a Declaration of Intent with the authorities.
The regular exporter then gives the declaration information to his suppliers.
The supplier can then issue an invoice to the regular exported
without taxes by noting the declaration of intent.

This commit adds support for this use case.
It is now possible to specificy a declaration of intent on sales orders and invoices.
They are also stated on the respective PDFs and electronic invoices.
A dedicated tax and fiscal position was added to be applied in case a declaration of intent is used.

Restrictions:
* The document currency has to be in EUR (more precisely the
  currency of the declaration of intent which defaults to EUR);
  else the declaration of intent cannot be applied
* There must not be any taxes on the document at all (else the invoice
  cannot be confirmed)

task-3794632