[ADD] l10n_it_edi_doi: Declaration of Intent support for l10n_it
Backport from 17.0: f9093fdbe76a4b9923735b335523925ac553eb01

In Italy, an operator who exports more than 10% of all his revenues is considered a regular exporter.
Regular exporters can purchase without paying VAT within certain limits.

For this they have to do a declaration of intent with the authorities.
The regular exporter then gives the declaration information to his suppliers.
The supplier can then issue an invoice to the regular exported
without taxes by noting the declaration of intent.

This commit adds support for the supplier side of this use case.
It is now possible to
- Create declarations of intent (of regular exporters we want to sell to)
- Use them on sales orders and invoices
- The not yet invoiced, invoiced and remaining amount of each
  declaration / SO / invoiced are tracked.
The declaration of intent information is also stated on the PDFs and electronic invoices.
A dedicated tax and fiscal position was added to be applied in case a declaration of intent is used.

Limitations:
- The document currency has to be in EUR
  (more precisely the currency of the declaration of intent which defaults to EUR);
  else the declaration of intent cannot be applied
- Lines containing the declaration of intent tax may not use any other tax
- When adding lines on a credit note linked to a sales order those lines do not count
  for the invoiced amount of the sales order. The sales order would have to be adapted.
  (It is not really a limitation because that is just how it works: the new lines are not linked to the sales order.)
- Documents can only use the declaration of intent if they lie inside the date range of the declaration and the declaration is in 'active' state.
  Exception: the invoiced amount is negative.

task-3794632