[FIX] l10n_test_pos_qr_payment: add missing journal on a pos.config
Currently the a pos.config used for the tests does not have a journal set.

In the code pos.config is assumed to have a `journal_id` though.
I.e. we take the company from this journal in function `_check_start_date`
of model 'pos.session' and call the function `_get_user_fiscal_lock_date` on it

In a different commit of this PR a call to `ensure_one` is added to
the lock date functions. The tests cause an error due to this.