[IMP] account: allow hashing moves in a locked period
After this commit moves can also be hashed when they are in a locked period.
We will no longer check for unhashed journal entries when trying to set a lock date (no more error).

Both Lock Dates and Hashing are mechanisms to make moves inalterable.
They do not really interfere with each other.

The new way is more user-friendly:
  * We do not force the user to hash everything before setting a lock date.
  * Hashing still possible after lock date.

task-4095970