[FIX] l10n_it_edi_doi: [?SQUASH?] split declaration downpayment amount
In 17.0 the downpayment lines are automatically split per tax and
the tax is set on the downpayment lines.

This commit introduces a dedicated logic just for the declaration of
intent tax: The downpayment amount (and downpayment line) is split
into 2: declaration of intent amount and other amounts. We also set
the special declaration of intent tax on the downpayment line.
This way the invoice will also have a dedicated line just for the
speical tax and has the special tax set.
This makes the backport (to 16.0) work.

[FIX] l10n_it_edi_doi: [SQUASH] on downpayment line do not map taxes with doi fiscal position

[FIX] l10n_it_edi_doi: fix downpayment amount when tax not price_included