[IMP] account: improve exception auditing
This commit improves the exception auditing in 2 ways
- The form view now displays the original date too (and not just the
  new / exceptional date)
- The button to show a restricted audit trail regarding changes during the
  exception was improved. See below for details

The exception form view provides a button to open the audit trail
restricted to messages that were created during the time the exception
was valid.
But this can also contain (many) changes that did not need an exception at all.

After this commit the audit trail is restricted further to accounting dates
in the "excepted period". The "excepted period" starts at the minimal
excepted date and ends at the the maximal original lock date.
E.g. when the Sale Lock Date was changed from 2024-02-01 to 2024-01-01
and the Purchase Lock Date was changed from 2024-03-01 to 2024-02-01
we regard the time period 2024-01-01 to 2024-03-01.

Messages are shown when
- the accounting date of the record is currently inside the "excepted period"
- the messages is regarding an accounting date change from or into the
  "excepted period"
This should find all moves that were changed during the exception
and only show moves for which we needed an exception.

related: task-3891414 (lock dates rework)