[IMP] account: audit trail; set record fields also when inactive
On each message there is a computed field for each supported model
containing the related record of that model.
E.g. if the message is about move x then x is put into field
`account_audit_log_move_id`.

But currently this only happens if the audit trail is activated
and supposed to restrict messages related to the record.

This is kind of unnecessary since there is also the field
`account_audit_log_activated` from which we can get this
information.

After this commit we fill the "related record fields" anyway.

TODO: check:
some domains etc may rely on the fact that is is not set